+++ mlcvs/tests/test_tica_linear.py
